Download latest version
Chocolatey website suggests that 19.9.0.22 is released, but this isn't correct: https://chocolatey.org/packages/citrix-workspace#versionhistory 19.9.0.21 is still the latest. I've updated the script to always get the latest (and output what it found) removed the hard coded stuff. I'm not able to test this out locally though as it won't load the HtmlAgilityPack DLL :(